requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# GraphQL
graphene-django==3.1.2
django-graphql-jwt==0.3.4

# API Documentation  
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0

# Production
gunicorn==21.2.0

# Development
django-extensions==3.2.3

# Image handling
Pillow==10.0.1

# Utilities
django-filter==23.3

# Sample data generation
Faker==19.6.2